C program that prints the size of various types on the computer it is compiled and run on, for both 32 and 64 bits

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,11 @@
  */
 int main(void)
 {
+
 	printf("Size of a char: %lu bytes\n", sizeof(char));
 	printf("Size of an int: %lu bytes\n", sizeof(int));
 	printf("Size of a long int: %lu bytes\n", sizeof(long));
-	printf("Size of a long long int: %lu bytes\n", sizeof(double));
+	printf("Size of a long long int: %lu bytes\n", sizeof(long long));
 	printf("Size of a float: %lu bytes\n", sizeof(float));
 
 	return (0);
